Create [EN] CodeHook.cpp
This code is intended to "hack" (modify) a function called originalFunction so that a call to that function directs execution to hookedFunction, which is our hacking function.

The process of hacking a function consists in modifying the executable code of the function in such a way that the call to the original function is replaced with the call of our hacking function. This way, when originalFunction is called elsewhere in the program code, instead of executing the original function code, the hookedFunction code will be executed.

In the code shown, we first allocate memory and copy the originalFunction code into that allocated memory. We then modify this copied code to replace the originalFunction call with a hookedFunction call. After making these modifications, we restore the memory page protection to read-only mode and perform the hacking function.

This way, when originalFunction is called somewhere else in the code, our hookedFunction will be executed. This gives us the opportunity to modify the behavior of the program in different ways, depending on what we want to achieve in the hacking function.

This is my 1st C++ program, so it may not be the best in terms of quality, if you have any objections to this code, write to me on this dc - https://discord.gg/XRhbCk2gaz